snap.c:183:45: error: implicit conversion loses integer precision: 'unsigned int' to 'uint8_t' (aka 'unsigned char') [-Werror,-Wconversion]

+++ test/format/snap.c
@@ -151,7 +151,7 @@ snap_track(TINFO *tinfo, thread_op op)
     SNAP_OPS *snap;
     TABLE *table;
     WT_ITEM *ip;
-    uint8_t mask;
+    u_int mask;
 
     table = tinfo->table;
